Changed NED2Subregion to accept zipped NED tiles
When not doing this in a huge batch, NED tiles usually stored compressed because they take up a lot of space.
If you want to do just a few subregions, this tool will unzip only the ones you need.
If you're doing them all at once, you'll unzip everything and leave the box unchecked but that should be in the documentation.