Moved mosaic function from PreHPCC (step 2) to Ned2Subregion
This is great for disk space management. There is no real reason to store
the unzipped "staged" files, just make the mosaic during the "staging" and delete the unzipped tiles when it's done.

This way, a user can run any number of these as a batch without running out of disk space due to redundant intermediates.

Signed-off-by: nicolejeansmith <nicole.j.smith@gmail.com>