Zonal Attribution for Raster Data now more robust, but slower
To avoid rare problems exemplified when I tried to use the tool with 100m lake buffer (I think it is something weird with the Permanent_Identifier string, but really a lot of weird stuff happened), the vector-to-raster conversion is now explicit. This slows down processing with finer resolution datasets but removes the problem I was having with unexplained bombing out.

Signed-off-by: nicolejeansmith <nicole.j.smith@gmail.com>