Added a try/except in zonal_tabarea to allow it to finish faster in some conditions
Before, when I had the problem with using Permanent Identifier as the zone field, I modified the tool to convert the zones to a zones raster that matches the value raster in a temporary geodatabase. Now, the tool ONLY does that if it fails to do Zonal Stats directly. This should let the tool run faster in the 75%+ of cases where that works just fine.

Signed-off-by: nicolejeansmith <nicole.j.smith@gmail.com>