Added a known bug list and outstanding features todo list
known and outstanding bugs.txt
outstanding feature todos.txt

Discovered an unnecessary warning thrown up when I ran a tool, the output is fine so it's okay to run the tool but I don't know why it issued the warning. This is low priority to me so I decided to add a bugs list, and then I thought to add the features I haven't gotten to, as well.